+++ src/main/scala/stellar/sdk/model/result/TransactionHistory.scala
@@ -39,6 +39,7 @@ object TransactionHistoryDeserializer extends ResponseParser[TransactionHistory]
         case "id" => MemoId((o \ "memo").extract[String].toLong)
         case "text" => MemoText((o \ "memo").extractOpt[String].getOrElse(""))
         case "hash" => MemoHash(base64((o \ "memo").extract[String]))
+        case "return" => MemoReturnHash(base64((o \ "memo").extract[String]))
       },
       signatures = (o \ "signatures").extract[Seq[String]],
       envelopeXDR = (o \ "envelope_xdr").extract[String],